refactor(app): refactor ClearUnavailableRobots in Advanced Settings
create a new component for ClearUnavailableRobots and add tests for that.

close RAUT-938 partially
