+++ app/src/organisms/AdvancedSettings/index.ts
@@ -1,3 +1,4 @@
+export * from './ClearUnavailableRobots'
 export * from './EnableDevTools'
 export * from './OT2AdvancedSettings'
 export * from './PreventRobotCaching'
