+++ app/src/pages/AppSettings/AdvancedSettings.tsx
@@ -14,51 +14,30 @@ import {
   SPACING,
   TYPOGRAPHY,
   DIRECTION_COLUMN,
-  useConditionalConfirm,
-  JUSTIFY_FLEX_END,
-  Btn,
-  AlertPrimaryButton,
-  DIRECTION_ROW,
 } from '@opentrons/components'
 
 import * as Config from '../../redux/config'
 import * as ProtocolAnalysis from '../../redux/protocol-analysis'
 import * as CustomLabware from '../../redux/custom-labware'
-import {
-  clearDiscoveryCache,
-  getReachableRobots,
-  getUnreachableRobots,
-} from '../../redux/discovery'
-import { LegacyModal } from '../../molecules/LegacyModal'
-import { Portal } from '../../App/portal'
 import { SelectOption } from '../../atoms/SelectField/Select'
 import { SelectField } from '../../atoms/SelectField'
-import { ERROR_TOAST, SUCCESS_TOAST } from '../../atoms/Toast'
 import {
   useTrackEvent,
   ANALYTICS_CHANGE_PATH_TO_PYTHON_DIRECTORY,
   ANALYTICS_CHANGE_CUSTOM_LABWARE_SOURCE_FOLDER,
 } from '../../redux/analytics'
-// import {
-//   getU2EAdapterDevice,
-//   getU2EWindowsDriverStatus,
-//   OUTDATED,
-// } from '../../redux/system-info'
 import { Divider } from '../../atoms/structure'
 import { TertiaryButton, ToggleButton } from '../../atoms/buttons'
 import { StyledText } from '../../atoms/text'
-// import { Banner } from '../../atoms/Banner'
-import { useToaster } from '../../organisms/ToasterOven'
 import {
+  ClearUnavailableRobots,
   EnableDevTools,
   OT2AdvancedSettings,
   PreventRobotCaching,
   U2EInformation,
 } from '../../organisms/AdvancedSettings'
 
-import type { Dispatch, State } from '../../redux/types'
-
-// const REALTEK_URL = 'https://www.realtek.com/en/'
+import type { Dispatch } from '../../redux/types'
 
 export function AdvancedSettings(): JSX.Element {
   const { t } = useTranslation(['app_settings', 'shared'])
@@ -77,39 +56,6 @@ export function AdvancedSettings(): JSX.Element {
   const pathToPythonInterpreter = useSelector(Config.getPathToPythonOverride)
 
   const dispatch = useDispatch<Dispatch>()
-  const { makeToast } = useToaster()
-  const reachableRobots = useSelector((state: State) =>
-    getReachableRobots(state)
-  )
-  const unreachableRobots = useSelector((state: State) =>
-    getUnreachableRobots(state)
-  )
-  const recentlySeenRobots = reachableRobots.filter(
-    robot => robot.healthStatus !== 'ok'
-  )
-  const isUnavailableRobots =
-    unreachableRobots.length > 0 || recentlySeenRobots.length > 0
-
-  const handleDeleteUnavailRobots = (): void => {
-    if (isUnavailableRobots) {
-      dispatch(clearDiscoveryCache())
-      makeToast(t('successfully_deleted_unavail_robots'), SUCCESS_TOAST)
-    } else {
-      makeToast(t('no_unavail_robots_to_clear'), ERROR_TOAST)
-    }
-  }
-
-  const {
-    confirm: confirmDeleteUnavailRobots,
-    showConfirmation: showConfirmDeleteUnavailRobots,
-    cancel: cancelExit,
-  } = useConditionalConfirm(handleDeleteUnavailRobots, true)
-
-  // const device = useSelector(getU2EAdapterDevice)
-  // const driverOutdated = useSelector((state: State) => {
-  //   const status = getU2EWindowsDriverStatus(state)
-  //   return status === OUTDATED
-  // })
 
   const toggleLabwareOffsetData = (): void => {
     dispatch(
@@ -163,42 +109,6 @@ export function AdvancedSettings(): JSX.Element {
           justifyContent={JUSTIFY_SPACE_BETWEEN}
           gridGap={SPACING.spacing40}
         >
-          {showConfirmDeleteUnavailRobots ? (
-            <Portal level="top">
-              <LegacyModal
-                type="warning"
-                title={t('clear_unavailable_robots')}
-                onClose={cancelExit}
-              >
-                <StyledText as="p">{t('clearing_cannot_be_undone')}</StyledText>
-                <Flex
-                  flexDirection={DIRECTION_ROW}
-                  paddingTop={SPACING.spacing32}
-                  justifyContent={JUSTIFY_FLEX_END}
-                >
-                  <Flex
-                    paddingRight={SPACING.spacing4}
-                    data-testid="AdvancedSettings_ConfirmClear_Cancel"
-                  >
-                    <Btn
-                      onClick={cancelExit}
-                      textTransform={TYPOGRAPHY.textTransformCapitalize}
-                      color={COLORS.blue50}
-                      fontWeight={TYPOGRAPHY.fontWeightSemiBold}
-                      marginRight={SPACING.spacing32}
-                    >
-                      {t('shared:cancel')}
-                    </Btn>
-                  </Flex>
-                  <Flex data-testid="AdvancedSettings_ConfirmClear_Proceed">
-                    <AlertPrimaryButton onClick={confirmDeleteUnavailRobots}>
-                      {t('clear_confirm')}
-                    </AlertPrimaryButton>
-                  </Flex>
-                </Flex>
-              </LegacyModal>
-            </Portal>
-          ) : null}
           <Flex flexDirection={DIRECTION_COLUMN}>
             <StyledText
               css={TYPOGRAPHY.h3SemiBold}
@@ -294,29 +204,7 @@ export function AdvancedSettings(): JSX.Element {
         <Divider marginY={SPACING.spacing24} />
         <PreventRobotCaching />
         <Divider marginY={SPACING.spacing24} />
-        <Flex
-          alignItems={ALIGN_CENTER}
-          justifyContent={JUSTIFY_SPACE_BETWEEN}
-          gridGap={SPACING.spacing40}
-        >
-          <Box>
-            <StyledText
-              css={TYPOGRAPHY.h3SemiBold}
-              paddingBottom={SPACING.spacing8}
-              id="AdvancedSettings_clearRobots"
-            >
-              {t('clear_unavail_robots')}
-            </StyledText>
-            <StyledText as="p">{t('clear_robots_description')}</StyledText>
-          </Box>
-          <TertiaryButton
-            marginLeft={SPACING_AUTO}
-            onClick={confirmDeleteUnavailRobots}
-            id="AdvancedSettings_clearUnavailableRobots"
-          >
-            {t('clear_robots_button')}
-          </TertiaryButton>
-        </Flex>
+        <ClearUnavailableRobots />
         <Divider marginY={SPACING.spacing24} />
         <Flex alignItems={ALIGN_CENTER} justifyContent={JUSTIFY_SPACE_BETWEEN}>
           <Box width="70%">
