+++ app/src/pages/AppSettings/__test__/AdvancedSettings.test.tsx
@@ -2,19 +2,10 @@ import * as React from 'react'
 import { MemoryRouter } from 'react-router-dom'
 import { resetAllWhenMocks } from 'jest-when'
 import { fireEvent, screen } from '@testing-library/react'
-import {
-  renderWithProviders,
-  useConditionalConfirm,
-} from '@opentrons/components'
-import {
-  getReachableRobots,
-  getUnreachableRobots,
-} from '../../../redux/discovery'
+
+import { renderWithProviders } from '@opentrons/components'
+
 import { i18n } from '../../../i18n'
-import {
-  mockReachableRobot,
-  mockUnreachableRobot,
-} from '../../../redux/discovery/__fixtures__'
 import {
   useTrackEvent,
   ANALYTICS_CHANGE_PATH_TO_PYTHON_DIRECTORY,
@@ -24,6 +15,7 @@ import * as CustomLabware from '../../../redux/custom-labware'
 import * as Config from '../../../redux/config'
 import * as ProtocolAnalysis from '../../../redux/protocol-analysis'
 import {
+  ClearUnavailableRobots,
   EnableDevTools,
   OT2AdvancedSettings,
   PreventRobotCaching,
@@ -53,15 +45,6 @@ const render = (): ReturnType<typeof renderWithProviders> => {
   )
 }
 
-const mockGetUnreachableRobots = getUnreachableRobots as jest.MockedFunction<
-  typeof getUnreachableRobots
->
-const mockGetReachableRobots = getReachableRobots as jest.MockedFunction<
-  typeof getReachableRobots
->
-const mockUseConditionalConfirm = useConditionalConfirm as jest.MockedFunction<
-  typeof useConditionalConfirm
->
 const getCustomLabwarePath = CustomLabware.getCustomLabwareDirectory as jest.MockedFunction<
   typeof CustomLabware.getCustomLabwareDirectory
 >
@@ -101,10 +84,11 @@ const mockEnableDevTools = EnableDevTools as jest.MockedFunction<
 const mockU2EInformation = U2EInformation as jest.MockedFunction<
   typeof U2EInformation
 >
+const mockClearUnavailableRobots = ClearUnavailableRobots as jest.MockedFunction<
+  typeof ClearUnavailableRobots
+>
 
 let mockTrackEvent: jest.Mock
-const mockConfirm = jest.fn()
-const mockCancel = jest.fn()
 
 describe('AdvancedSettings', () => {
   beforeEach(() => {
@@ -119,17 +103,13 @@ describe('AdvancedSettings', () => {
       { label: 'Beta', value: 'beta' },
       { label: 'Alpha', value: 'alpha' },
     ])
-    mockGetUnreachableRobots.mockReturnValue([mockUnreachableRobot])
-    mockGetReachableRobots.mockReturnValue([mockReachableRobot])
-    mockUseConditionalConfirm.mockReturnValue({
-      confirm: mockConfirm,
-      showConfirmation: true,
-      cancel: mockCancel,
-    })
     mockPreventRobotCaching.mockReturnValue(<div>mock PreventRobotCaching</div>)
     mockOT2AdvancedSettings.mockReturnValue(<div>mock OT2AdvancedSettings</div>)
     mockEnableDevTools.mockReturnValue(<div>mock EnableDevTools</div>)
     mockU2EInformation.mockReturnValue(<div>mock U2EInformation</div>)
+    mockClearUnavailableRobots.mockReturnValue(
+      <div>mock ClearUnavailableRobots</div>
+    )
   })
 
   afterEach(() => {
@@ -141,7 +121,6 @@ describe('AdvancedSettings', () => {
     const [{ getByText }] = render()
     getByText('Update Channel')
     getByText('Additional Custom Labware Source Folder')
-    getByText('Clear Unavailable Robots')
   })
 
   it('renders the update channel combobox and section', () => {
@@ -260,29 +239,9 @@ describe('AdvancedSettings', () => {
     expect(mockOpenPythonInterpreterDirectory).toHaveBeenCalled()
   })
 
-  it('renders the clear unavailable robots section', () => {
-    const [{ getByText, getByRole }] = render()
-    getByText(
-      'Clear the list of unavailable robots on the Devices page. This action cannot be undone.'
-    )
-    const btn = getByRole('button', {
-      name: 'Clear unavailable robots list',
-    })
-    fireEvent.click(btn)
-    getByText('Clear unavailable robots?')
-    getByText(
-      'Clearing the list of unavailable robots on the Devices page cannot be undone.'
-    )
-    const closeBtn = getByRole('button', {
-      name: 'cancel',
-    })
-    const proceedBtn = getByRole('button', {
-      name: 'Clear unavailable robots',
-    })
-    fireEvent.click(closeBtn)
-    expect(mockCancel).toHaveBeenCalled()
-    fireEvent.click(proceedBtn)
-    expect(mockConfirm).toHaveBeenCalled()
+  it('should render mock clear unavailable robots section', () => {
+    render()
+    screen.getByText('mock ClearUnavailableRobots')
   })
 
   it('should render mock developer tools section', () => {
